Refactoring, moved analyze package
