+++ packages/backend/src/model/analyze/average.js
