+++ packages/backend/src/model/analyze/minimum.js
