+++ packages/backend/src/model/analyze/reduceNested.js
@@ -1,4 +1,4 @@
-const { isObject } = require("../../../utils");
+const { isObject } = require("../../utils");
 
 module.exports = process => ({
   reducer,
