+++ packages/backend/src/model/analyze/reduceNestedCount.js
